requirements.txt
annotated-types==0.7.0
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
fastapi==0.115.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.0
matplotlib==3.9.2
motor==3.6.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pymongo==4.9.2
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
rubicon-objc==0.4.9
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.7
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.31.0
